Place add_user_to_org in explicitly required module
Chef 12.1+ contains a Chef::Org class that was derived from this class
but which is missing the add_user_to_org method. This is part of
ongoing work to move knife-opc into core chef.  However, this causes
undefined method errors since require 'chef/org' no longer loads our
version of this class.  Users of Chef 12.1 can now use:

   require 'chef/org/group_operations'

to add this method to the Chef::Org object.

Fixes #28
