+++ lib/chef/knife/opc_org_create.rb
@@ -35,6 +35,7 @@ module Opc
 
     deps do
       require 'chef/org'
+      require 'chef/org/group_operations'
     end
 
     def run
