+++ lib/chef/knife/opc_org_user_add.rb
@@ -29,6 +29,7 @@ module Opc
 
     deps do
       require 'chef/org'
+      require 'chef/org/group_operations'
     end
 
     def run
